requirements

+++ src/app.py
@@ -1,51 +1,53 @@
-import os.path
+import os
 import uuid
 import streamlit as st
+import matplotlib.pyplot as plt
+from googleapiclient.discovery import build
+from googleapiclient.http import MediaIoBaseDownload
 from google_auth_oauthlib.flow import InstalledAppFlow
 from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
-from googleapiclient.discovery import build
-from googleapiclient.http import MediaIoBaseDownload
 from helper import extract_data_analysis, get_pdf_paths, read_uploaded_file
 from database import AnalyzeDatabase
 from ai import GroqClient
 from models.resum import Resum
 from models.file import File
-import matplotlib.pyplot as plt
 
-# Defina o escopo de acesso que você precisa
+# Define o escopo de acesso
 SCOPES = [
     "https://www.googleapis.com/auth/drive.file",
     "https://www.googleapis.com/auth/drive.readonly",
     "https://www.googleapis.com/auth/drive.metadata.readonly"
 ]
 
-# Inicialize a variável de credenciais
+# Inicializa as credenciais
 creds = None
 
-# Verifica se o arquivo token.json existe no diretório atual
-if os.path.exists('token.json'):
-    # Se existir, carrega as credenciais a partir do arquivo token.json
-    creds = Credentials.from_authorized_user_file('token.json', SCOPES)
-    st.success('Credenciais carregadas do arquivo token.json.')
+# Verifique se o arquivo credentials.json existe
+credentials_path = 'analyze/credentials.json'  # Altere o caminho conforme necessário
+if not os.path.exists(credentials_path):
+    raise FileNotFoundError('O arquivo credentials.json não foi encontrado. Por favor, crie-o manualmente no Google Cloud Console.')
 
-# Verifica se as credenciais não existem ou são inválidas/expiradas
+# Verifique se o arquivo token.json existe
+if os.path.exists('token.json'):
+    creds = Credentials.from_authorized_user_file('token.json', SCOPES)
+    print('Credenciais carregadas do arquivo token.json.')
+
+# Verifique se as credenciais não existem ou são inválidas/expiradas
 if not creds or not creds.valid:
-    # Se as credenciais existirem mas estiverem expiradas e puderem ser renovadas
     if creds and creds.expired and creds.refresh_token:
         creds.refresh(Request())
-        st.success('Credenciais renovadas com sucesso.')
+        print('Credenciais renovadas com sucesso.')
     else:
-        # Se não houver credenciais válidas, inicia o fluxo de autorização do OAuth
-        st.warning('Iniciando o fluxo de autorização do OAuth...')
-        flow = InstalledAppFlow.from_client_secrets_file('credentials.json', SCOPES)
+        print('Iniciando o fluxo de autorização do OAuth...')
+        flow = InstalledAppFlow.from_client_secrets_file(credentials_path, SCOPES)
         creds = flow.run_local_server(port=0)  # Executa o servidor local para autorização
-        st.success('Autorização concluída.')
+        print('Autorização concluída.')
 
     # Salva as credenciais renovadas ou novas no arquivo token.json
     with open('token.json', 'w') as token:
         token.write(creds.to_json())
-        st.success('Credenciais salvas no arquivo token.json.')
+        print('Credenciais salvas no arquivo token.json.')
 
 # Construa o serviço da API Google Drive
 service = build('drive', 'v3', credentials=creds)
@@ -53,30 +55,31 @@ service = build('drive', 'v3', credentials=creds)
 # ID da pasta que você deseja listar os arquivos
 folder_id = '1bnvaXQq7s1gJyh4w-b6QnW4fT-C6mIzK'
 
-# Listar arquivos na pasta especificada pelo folder_id
+# Lista arquivos na pasta especificada pelo folder_id
 results = service.files().list(
     q=f"'{folder_id}' in parents", fields="files(id, name)"
 ).execute()
 
-# Obter a lista de arquivos
+# Obtém a lista de arquivos
 files = results.get('files', [])
 
+# Download de cada arquivo no drive
 if not files:
-    st.warning("⚠️ Nenhum currículo encontrado na pasta do Google Drive.")
+    raise FileNotFoundError('Nenhum arquivo encontrado.')
 else:
+    print('Files:')
     for file in files:
-        st.write(f"Baixando: {file['name']} ({file['id']})")
+        print(f"{file['name']} ({file['id']})")
 
         # Download de cada arquivo no drive
         request = service.files().get_media(fileId=file['id'])
         file_path = f"cv-analyzer/analyze/drive/curriculos/{file['name']}"  # Define o caminho de onde salvar o arquivo
-        
         with open(file_path, 'wb') as f:
             downloader = MediaIoBaseDownload(f, request)
             done = False
             while not done:
                 status, done = downloader.next_chunk()
-                st.write(f"Download {int(status.progress() * 100)}% concluído.")
+                print(f"Download {int(status.progress() * 100)}%.")
 
 # Inicializa a base de dados e a IA
 database = AnalyzeDatabase()
@@ -92,7 +95,7 @@ st.set_page_config(layout="wide", page_title="Análise de Currículos", page_ico
 st.title("🔍 Análise de Currículos para a Vaga: **Gestor Comercial de B2B**")
 
 # Pega o caminho dos currículos no diretório especificado
-cv_paths = get_pdf_paths(directory='cv-analyzer/analyze/drive/curriculos')  # Use um caminho relativo
+cv_paths = get_pdf_paths(directory='cv-analyzer/analyze/drive/curriculos')
 
 # Lista para armazenar os scores dos candidatos
 scores = []
